Disposable is now a base class which auto-registers
This is memory overhaul #2. In this CL, we remove the Box class -
it was causing more trouble than its worth. Instead, we redesigned
Disposables so they auto-register on construction and callers must
remember to explicitly dispose them later.

This does, unfortunately, allow for code to be called on objects that
should already have been disposed, but in practice that probably wasn't
a major problem in most cases anyway.

We also added a `Lifetimes` interface, removing the Disposable interface
from the GameState and App classes, instead providing disposables
through the Init and Update contexts.